chore: move code reponsible for transforming objects to gql_generated types to separate package
