+++ pkg/extensions/search/resolver.go
@@ -6,11 +6,7 @@ package search
 
 import (
 	"context"
-	"encoding/json"
-	"fmt"
-	"strconv"
 	"strings"
-	"time"
 
 	"github.com/99designs/gqlgen/graphql"
 	glob "github.com/bmatcuk/doublestar/v4"            //nolint:gci
@@ -21,6 +17,7 @@ import (
 	"github.com/vektah/gqlparser/v2/gqlerror"
 	zerr "zotregistry.io/zot/errors"
 	"zotregistry.io/zot/pkg/extensions/search/common"
+	"zotregistry.io/zot/pkg/extensions/search/convert"
 	cveinfo "zotregistry.io/zot/pkg/extensions/search/cve"
 	digestinfo "zotregistry.io/zot/pkg/extensions/search/digest"
 	"zotregistry.io/zot/pkg/extensions/search/gql_generated"
@@ -82,7 +79,7 @@ func (r *queryResolver) getImageListForDigest(repoList []string, digest string) 
 			}
 
 			isSigned := olu.CheckManifestSignature(repo, imageInfo.Digest)
-			imageInfo := BuildImageInfo(repo, imageInfo.Tag, imageInfo.Digest,
+			imageInfo := convert.BuildImageInfo(repo, imageInfo.Tag, imageInfo.Digest,
 				imageInfo.Manifest, imageConfig, isSigned)
 
 			imgResultForDigest = append(imgResultForDigest, imageInfo)
@@ -116,7 +113,7 @@ func getImageSummary(ctx context.Context, repo, tag string, repoDB repodb.RepoDB
 		manifestDigest: manifestMeta,
 	}
 
-	imageSummaries := RepoMeta2ImageSummaries(ctx, repoMeta, manifestMetaMap, cveInfo)
+	imageSummaries := convert.RepoMeta2ImageSummaries(ctx, repoMeta, manifestMetaMap, cveInfo)
 
 	return imageSummaries[0], nil
 }
@@ -148,7 +145,7 @@ func repoListWithNewestImage(
 	}
 
 	for _, repoMeta := range reposMeta {
-		repoSummary := RepoMeta2RepoSummary(ctx, repoMeta, manifestMetaMap, cveInfo)
+		repoSummary := convert.RepoMeta2RepoSummary(ctx, repoMeta, manifestMetaMap, cveInfo)
 		repos = append(repos, repoSummary)
 	}
 
@@ -197,7 +194,7 @@ func globalSearch(ctx context.Context, query string, repoDB repodb.RepoDB, filte
 		}
 
 		for _, repoMeta := range reposMeta {
-			repoSummary := RepoMeta2RepoSummary(ctx, repoMeta, manifestMetaMap, cveInfo)
+			repoSummary := convert.RepoMeta2RepoSummary(ctx, repoMeta, manifestMetaMap, cveInfo)
 
 			*repoSummary.Score = calculateImageMatchingScore(repoMeta.Name, strings.Index(repoMeta.Name, query))
 			repos = append(repos, repoSummary)
@@ -217,7 +214,7 @@ func globalSearch(ctx context.Context, query string, repoDB repodb.RepoDB, filte
 		}
 
 		for _, repoMeta := range reposMeta {
-			imageSummaries := RepoMeta2ImageSummaries(ctx, repoMeta, manifestMetaMap, cveInfo)
+			imageSummaries := convert.RepoMeta2ImageSummaries(ctx, repoMeta, manifestMetaMap, cveInfo)
 
 			images = append(images, imageSummaries...)
 		}
@@ -259,7 +256,7 @@ func expandedRepoInfo(ctx context.Context, repo string, repoDB repodb.RepoDB, cv
 		manifestMetaMap[digest] = manifestMeta
 	}
 
-	repoSummary, imageSummaries := RepoMeta2ExpandedRepoInfo(ctx, repoMeta, manifestMetaMap, cveInfo)
+	repoSummary, imageSummaries := convert.RepoMeta2ExpandedRepoInfo(ctx, repoMeta, manifestMetaMap, cveInfo)
 
 	return &gql_generated.RepoInfo{Summary: repoSummary, Images: imageSummaries}, nil
 }
@@ -272,540 +269,8 @@ func safeDerefferencing[T any](pointer *T, defaultVal T) T {
 	return defaultVal
 }
 
-func RepoMeta2ImageSummaries(ctx context.Context, repoMeta repodb.RepoMetadata,
-	manifestMetaMap map[string]repodb.ManifestMetadata, cveInfo cveinfo.CveInfo,
-) []*gql_generated.ImageSummary {
-	imageSummaries := make([]*gql_generated.ImageSummary, 0, len(repoMeta.Tags))
-
-	for tag, manifestDigest := range repoMeta.Tags {
-		var manifestContent ispec.Manifest
-
-		err := json.Unmarshal(manifestMetaMap[manifestDigest].ManifestBlob, &manifestContent)
-		if err != nil {
-			graphql.AddError(ctx, gqlerror.Errorf("can't unmarshal manifest blob for image: %s:%s, "+
-				"manifest digest: %s, error: %s", repoMeta.Name, tag, manifestDigest, err.Error()))
-
-			continue
-		}
-
-		var configContent ispec.Image
-
-		err = json.Unmarshal(manifestMetaMap[manifestDigest].ConfigBlob, &configContent)
-		if err != nil {
-			graphql.AddError(ctx, gqlerror.Errorf("can't unmarshal config blob for image: %s:%s, "+
-				"manifest digest: %s, error: %s", repoMeta.Name, tag, manifestDigest, err.Error()))
-
-			continue
-		}
-
-		imageCveSummary := cveinfo.ImageCVESummary{}
-		// Check if vulnerability scanning is disabled
-		if cveInfo != nil {
-			imageName := fmt.Sprintf("%s:%s", repoMeta.Name, tag)
-			imageCveSummary, err = cveInfo.GetCVESummaryForImage(imageName)
-
-			if err != nil {
-				// Log the error, but we should still include the manifest in results
-				graphql.AddError(ctx, gqlerror.Errorf("unable to run vulnerability scan on tag %s in repo %s: "+
-					"manifest digest: %s, error: %s", tag, repoMeta.Name, manifestDigest, err.Error()))
-			}
-		}
-
-		imgSize := int64(0)
-		imgSize += manifestContent.Config.Size
-		imgSize += int64(len(manifestMetaMap[manifestDigest].ManifestBlob))
-
-		for _, layer := range manifestContent.Layers {
-			imgSize += layer.Size
-		}
-
-		var (
-			repoName         = repoMeta.Name
-			tag              = tag
-			manifestDigest   = manifestDigest
-			configDigest     = manifestContent.Config.Digest.String()
-			imageLastUpdated = common.GetImageLastUpdated(configContent)
-			isSigned         = imageHasSignatures(manifestMetaMap[manifestDigest].Signatures)
-			imageSize        = strconv.FormatInt(imgSize, 10)
-			os               = configContent.OS
-			arch             = configContent.Architecture
-			osArch           = gql_generated.OsArch{Os: &os, Arch: &arch}
-			downloadCount    = manifestMetaMap[manifestDigest].DownloadCount
-		)
-
-		annotations := common.GetAnnotations(manifestContent.Annotations, configContent.Config.Labels)
-
-		historyEntries, err := getAllHistory(manifestContent, configContent)
-		if err != nil {
-			graphql.AddError(ctx, gqlerror.Errorf("error generating history on tag %s in repo %s: "+
-				"manifest digest: %s, error: %s", tag, repoMeta.Name, manifestDigest, err.Error()))
-		}
-
-		imageSummary := gql_generated.ImageSummary{
-			RepoName:      &repoName,
-			Tag:           &tag,
-			Digest:        &manifestDigest,
-			ConfigDigest:  &configDigest,
-			LastUpdated:   &imageLastUpdated,
-			IsSigned:      &isSigned,
-			Size:          &imageSize,
-			Platform:      &osArch,
-			Vendor:        &annotations.Vendor,
-			DownloadCount: &downloadCount,
-			Layers:        getLayersSummaries(manifestContent),
-			Description:   &annotations.Description,
-			Title:         &annotations.Title,
-			Documentation: &annotations.Documentation,
-			Licenses:      &annotations.Licenses,
-			Labels:        &annotations.Labels,
-			Source:        &annotations.Source,
-			Logo:          &annotations.Logo,
-			History:       historyEntries,
-			Vulnerabilities: &gql_generated.ImageVulnerabilitySummary{
-				MaxSeverity: &imageCveSummary.MaxSeverity,
-				Count:       &imageCveSummary.Count,
-			},
-		}
-
-		imageSummaries = append(imageSummaries, &imageSummary)
-	}
-
-	return imageSummaries
-}
-
-func RepoMeta2ExpandedRepoInfo(ctx context.Context, repoMeta repodb.RepoMetadata,
-	manifestMetaMap map[string]repodb.ManifestMetadata, cveInfo cveinfo.CveInfo,
-) (*gql_generated.RepoSummary, []*gql_generated.ImageSummary) {
-	var (
-		repoLastUpdatedTimestamp = time.Time{}
-		repoPlatformsSet         = map[string]*gql_generated.OsArch{}
-		repoVendorsSet           = map[string]bool{}
-		lastUpdatedImageSummary  *gql_generated.ImageSummary
-		repoStarCount            = repoMeta.Stars
-		isBookmarked             = false
-		isStarred                = false
-		repoDownloadCount        = 0
-		repoName                 = repoMeta.Name
-
-		// map used to keep track of all blobs of a repo without dublicates as
-		// some images may have the same layers
-		repoBlob2Size = make(map[string]int64, 10)
-
-		// made up of all manifests, configs and image layers
-		size = int64(0)
-
-		imageSummaries = make([]*gql_generated.ImageSummary, 0, len(repoMeta.Tags))
-	)
-
-	for tag, manifestDigest := range repoMeta.Tags {
-		var manifestContent ispec.Manifest
-
-		err := json.Unmarshal(manifestMetaMap[manifestDigest].ManifestBlob, &manifestContent)
-		if err != nil {
-			graphql.AddError(ctx, gqlerror.Errorf("can't unmarshal manifest blob for image: %s:%s, manifest digest: %s, "+
-				"error: %s", repoMeta.Name, tag, manifestDigest, err.Error()))
-
-			continue
-		}
-
-		var configContent ispec.Image
-
-		err = json.Unmarshal(manifestMetaMap[manifestDigest].ConfigBlob, &configContent)
-		if err != nil {
-			graphql.AddError(ctx, gqlerror.Errorf("can't unmarshal config blob for image: %s:%s, manifest digest: %s, error: %s",
-				repoMeta.Name, tag, manifestDigest, err.Error()))
-
-			continue
-		}
-
-		imageCveSummary := cveinfo.ImageCVESummary{}
-		// Check if vulnerability scanning is disabled
-		if cveInfo != nil {
-			imageName := fmt.Sprintf("%s:%s", repoMeta.Name, tag)
-			imageCveSummary, err = cveInfo.GetCVESummaryForImage(imageName)
-
-			if err != nil {
-				// Log the error, but we should still include the manifest in results
-				graphql.AddError(ctx, gqlerror.Errorf("unable to run vulnerability scan on tag %s in repo %s: "+
-					"manifest digest: %s, error: %s", tag, repoMeta.Name, manifestDigest, err.Error()))
-			}
-		}
-
-		var (
-			tag              = tag
-			isSigned         = len(manifestMetaMap[manifestDigest].Signatures) > 0
-			configDigest     = manifestContent.Config.Digest.String()
-			configSize       = manifestContent.Config.Size
-			opSys            = configContent.OS
-			arch             = configContent.Architecture
-			osArch           = gql_generated.OsArch{Os: &opSys, Arch: &arch}
-			imageLastUpdated = common.GetImageLastUpdated(configContent)
-			downloadCount    = manifestMetaMap[manifestDigest].DownloadCount
-			manifestDigest   = manifestDigest
-
-			size = updateRepoBlobsMap(
-				manifestDigest, int64(len(manifestMetaMap[manifestDigest].ManifestBlob)),
-				configDigest, configSize,
-				manifestContent.Layers,
-				repoBlob2Size)
-			imageSize = strconv.FormatInt(size, 10)
-		)
-
-		annotations := common.GetAnnotations(manifestContent.Annotations, configContent.Config.Labels)
-
-		imageSummary := gql_generated.ImageSummary{
-			RepoName:      &repoName,
-			Tag:           &tag,
-			Digest:        &manifestDigest,
-			ConfigDigest:  &configDigest,
-			LastUpdated:   &imageLastUpdated,
-			IsSigned:      &isSigned,
-			Size:          &imageSize,
-			Platform:      &osArch,
-			Vendor:        &annotations.Vendor,
-			DownloadCount: &downloadCount,
-			Layers:        getLayersSummaries(manifestContent),
-			Description:   &annotations.Description,
-			Title:         &annotations.Title,
-			Documentation: &annotations.Documentation,
-			Licenses:      &annotations.Licenses,
-			Labels:        &annotations.Labels,
-			Source:        &annotations.Source,
-			Vulnerabilities: &gql_generated.ImageVulnerabilitySummary{
-				MaxSeverity: &imageCveSummary.MaxSeverity,
-				Count:       &imageCveSummary.Count,
-			},
-		}
-
-		imageSummaries = append(imageSummaries, &imageSummary)
-
-		if annotations.Vendor != "" {
-			repoVendorsSet[annotations.Vendor] = true
-		}
-
-		if opSys != "" || arch != "" {
-			osArchString := strings.TrimSpace(fmt.Sprintf("%s %s", opSys, arch))
-			repoPlatformsSet[osArchString] = &gql_generated.OsArch{Os: &opSys, Arch: &arch}
-		}
-
-		if repoLastUpdatedTimestamp.Equal(time.Time{}) {
-			// initialize with first time value
-			repoLastUpdatedTimestamp = imageLastUpdated
-			lastUpdatedImageSummary = &imageSummary
-		} else if repoLastUpdatedTimestamp.Before(imageLastUpdated) {
-			repoLastUpdatedTimestamp = imageLastUpdated
-			lastUpdatedImageSummary = &imageSummary
-		}
-
-		repoDownloadCount += manifestMetaMap[manifestDigest].DownloadCount
-	}
-
-	// calculate repo size = sum all manifest, config and layer blobs sizes
-	for _, blobSize := range repoBlob2Size {
-		size += blobSize
-	}
-
-	repoSize := strconv.FormatInt(size, 10)
-	score := 0
-
-	repoPlatforms := make([]*gql_generated.OsArch, 0, len(repoPlatformsSet))
-	for _, osArch := range repoPlatformsSet {
-		repoPlatforms = append(repoPlatforms, osArch)
-	}
-
-	repoVendors := make([]*string, 0, len(repoVendorsSet))
-
-	for vendor := range repoVendorsSet {
-		vendor := vendor
-		repoVendors = append(repoVendors, &vendor)
-	}
-
-	summary := &gql_generated.RepoSummary{
-		Name:          &repoName,
-		LastUpdated:   &repoLastUpdatedTimestamp,
-		Size:          &repoSize,
-		Platforms:     repoPlatforms,
-		Vendors:       repoVendors,
-		Score:         &score,
-		NewestImage:   lastUpdatedImageSummary,
-		DownloadCount: &repoDownloadCount,
-		StarCount:     &repoStarCount,
-		IsBookmarked:  &isBookmarked,
-		IsStarred:     &isStarred,
-	}
-
-	return summary, imageSummaries
-}
-
-func getLayersSummaries(manifestContent ispec.Manifest) []*gql_generated.LayerSummary {
-	layers := make([]*gql_generated.LayerSummary, 0, len(manifestContent.Layers))
-
-	for _, layer := range manifestContent.Layers {
-		size := strconv.FormatInt(layer.Size, 10)
-		digest := layer.Digest.String()
-
-		layers = append(layers, &gql_generated.LayerSummary{
-			Size:   &size,
-			Digest: &digest,
-		})
-	}
-
-	return layers
-}
-
-func getAllHistory(manifestContent ispec.Manifest, configContent ispec.Image) (
-	[]*gql_generated.LayerHistory, error,
-) {
-	allHistory := []*gql_generated.LayerHistory{}
-	layerSummaries := getLayersSummaries(manifestContent)
-
-	history := configContent.History
-	if len(history) == 0 {
-		// We don't have any image history metadata
-		// let's make due with just the layer metadata
-		for _, layer := range layerSummaries {
-			allHistory = append(allHistory, &gql_generated.LayerHistory{
-				Layer:              layer,
-				HistoryDescription: &gql_generated.HistoryDescription{},
-			})
-		}
-
-		return allHistory, nil
-	}
-
-	// Iterator over manifest layers
-	var layersIterator int
-	// Since we are appending pointers, it is important to iterate with an index over slice
-	for i := range history {
-		allHistory = append(allHistory, &gql_generated.LayerHistory{
-			HistoryDescription: &gql_generated.HistoryDescription{
-				Created:    history[i].Created,
-				CreatedBy:  &history[i].CreatedBy,
-				Author:     &history[i].Author,
-				Comment:    &history[i].Comment,
-				EmptyLayer: &history[i].EmptyLayer,
-			},
-		})
-
-		if history[i].EmptyLayer {
-			continue
-		}
-
-		if layersIterator+1 > len(manifestContent.Layers) {
-			return allHistory, zerr.ErrBadLayerCount
-		}
-
-		allHistory[i].Layer = layerSummaries[layersIterator]
-
-		layersIterator++
-	}
-
-	return allHistory, nil
-}
-
-func RepoMeta2RepoSummary(ctx context.Context, repoMeta repodb.RepoMetadata,
-	manifestMetaMap map[string]repodb.ManifestMetadata, cveInfo cveinfo.CveInfo,
-) *gql_generated.RepoSummary {
-	var (
-		repoLastUpdatedTimestamp = time.Time{}
-		repoPlatformsSet         = map[string]*gql_generated.OsArch{}
-		repoVendorsSet           = map[string]bool{}
-		lastUpdatedImageSummary  *gql_generated.ImageSummary
-		repoStarCount            = repoMeta.Stars
-		isBookmarked             = false
-		isStarred                = false
-		repoDownloadCount        = 0
-		repoName                 = repoMeta.Name
-
-		// map used to keep track of all blobs of a repo without dublicates as
-		// some images may have the same layers
-		repoBlob2Size = make(map[string]int64, 10)
-
-		// made up of all manifests, configs and image layers
-		size = int64(0)
-	)
-
-	for tag, manifestDigest := range repoMeta.Tags {
-		var manifestContent ispec.Manifest
-
-		err := json.Unmarshal(manifestMetaMap[manifestDigest].ManifestBlob, &manifestContent)
-		if err != nil {
-			graphql.AddError(ctx, gqlerror.Errorf("can't unmarshal manifest blob for image: %s:%s, manifest digest: %s, "+
-				"error: %s", repoMeta.Name, tag, manifestDigest, err.Error()))
-
-			continue
-		}
-
-		var configContent ispec.Image
-
-		err = json.Unmarshal(manifestMetaMap[manifestDigest].ConfigBlob, &configContent)
-		if err != nil {
-			graphql.AddError(ctx, gqlerror.Errorf("can't unmarshal config blob for image: %s:%s, manifest digest: %s, error: %s",
-				repoMeta.Name, tag, manifestDigest, err.Error()))
-
-			continue
-		}
-
-		imageCveSummary := cveinfo.ImageCVESummary{}
-		// Check if vulnerability scanning is disabled
-		if cveInfo != nil {
-			imageName := fmt.Sprintf("%s:%s", repoMeta.Name, tag)
-			imageCveSummary, err = cveInfo.GetCVESummaryForImage(imageName)
-
-			if err != nil {
-				// Log the error, but we should still include the manifest in results
-				graphql.AddError(ctx, gqlerror.Errorf("unable to run vulnerability scan on tag %s in repo %s: "+
-					"manifest digest: %s, error: %s", tag, repoMeta.Name, manifestDigest, err.Error()))
-			}
-		}
-
-		var (
-			tag              = tag
-			isSigned         = len(manifestMetaMap[manifestDigest].Signatures) > 0
-			configDigest     = manifestContent.Config.Digest.String()
-			configSize       = manifestContent.Config.Size
-			opSys            = configContent.OS
-			arch             = configContent.Architecture
-			osArch           = gql_generated.OsArch{Os: &opSys, Arch: &arch}
-			imageLastUpdated = common.GetImageLastUpdated(configContent)
-			downloadCount    = manifestMetaMap[manifestDigest].DownloadCount
-			manifestDigest   = manifestDigest
-
-			size = updateRepoBlobsMap(
-				manifestDigest, int64(len(manifestMetaMap[manifestDigest].ManifestBlob)),
-				configDigest, configSize,
-				manifestContent.Layers,
-				repoBlob2Size)
-			imageSize = strconv.FormatInt(size, 10)
-		)
-
-		annotations := common.GetAnnotations(manifestContent.Annotations, configContent.Config.Labels)
-
-		historyEntries, err := getAllHistory(manifestContent, configContent)
-		if err != nil {
-			graphql.AddError(ctx, gqlerror.Errorf("error generating history on tag %s in repo %s: "+
-				"manifest digest: %s, error: %s", tag, repoMeta.Name, manifestDigest, err.Error()))
-		}
-
-		imageSummary := gql_generated.ImageSummary{
-			RepoName:      &repoName,
-			Tag:           &tag,
-			Digest:        &manifestDigest,
-			ConfigDigest:  &configDigest,
-			LastUpdated:   &imageLastUpdated,
-			IsSigned:      &isSigned,
-			Size:          &imageSize,
-			Platform:      &osArch,
-			Vendor:        &annotations.Vendor,
-			DownloadCount: &downloadCount,
-			Layers:        getLayersSummaries(manifestContent),
-			Description:   &annotations.Description,
-			Title:         &annotations.Title,
-			Documentation: &annotations.Documentation,
-			Licenses:      &annotations.Licenses,
-			Labels:        &annotations.Labels,
-			Source:        &annotations.Source,
-			Logo:          &annotations.Logo,
-			History:       historyEntries,
-			Vulnerabilities: &gql_generated.ImageVulnerabilitySummary{
-				MaxSeverity: &imageCveSummary.MaxSeverity,
-				Count:       &imageCveSummary.Count,
-			},
-		}
-
-		if annotations.Vendor != "" {
-			repoVendorsSet[annotations.Vendor] = true
-		}
-
-		if opSys != "" || arch != "" {
-			osArchString := strings.TrimSpace(fmt.Sprintf("%s %s", opSys, arch))
-			repoPlatformsSet[osArchString] = &gql_generated.OsArch{Os: &opSys, Arch: &arch}
-		}
-
-		if repoLastUpdatedTimestamp.Equal(time.Time{}) {
-			// initialize with first time value
-			repoLastUpdatedTimestamp = imageLastUpdated
-			lastUpdatedImageSummary = &imageSummary
-		} else if repoLastUpdatedTimestamp.Before(imageLastUpdated) {
-			repoLastUpdatedTimestamp = imageLastUpdated
-			lastUpdatedImageSummary = &imageSummary
-		}
-
-		repoDownloadCount += manifestMetaMap[manifestDigest].DownloadCount
-	}
-
-	// calculate repo size = sum all manifest, config and layer blobs sizes
-	for _, blobSize := range repoBlob2Size {
-		size += blobSize
-	}
-
-	repoSize := strconv.FormatInt(size, 10)
-	score := 0
-
-	repoPlatforms := make([]*gql_generated.OsArch, 0, len(repoPlatformsSet))
-	for _, osArch := range repoPlatformsSet {
-		repoPlatforms = append(repoPlatforms, osArch)
-	}
-
-	repoVendors := make([]*string, 0, len(repoVendorsSet))
-
-	for vendor := range repoVendorsSet {
-		vendor := vendor
-		repoVendors = append(repoVendors, &vendor)
-	}
-
-	return &gql_generated.RepoSummary{
-		Name:          &repoName,
-		LastUpdated:   &repoLastUpdatedTimestamp,
-		Size:          &repoSize,
-		Platforms:     repoPlatforms,
-		Vendors:       repoVendors,
-		Score:         &score,
-		NewestImage:   lastUpdatedImageSummary,
-		DownloadCount: &repoDownloadCount,
-		StarCount:     &repoStarCount,
-		IsBookmarked:  &isBookmarked,
-		IsStarred:     &isStarred,
-	}
-}
-
-func imageHasSignatures(signatures map[string][]string) bool {
-	// (sigType, signatures)
-	for _, sigs := range signatures {
-		if len(sigs) > 0 {
-			return true
-		}
-	}
-
-	return false
-}
-
 func searchingForRepos(query string) bool {
 	return !strings.Contains(query, ":")
-}
-
-// updateRepoBlobsMap adds all the image blobs and their respective size to the repo blobs map
-// and returnes the total size of the image.
-func updateRepoBlobsMap(manifestDigest string, manifestSize int64, configDigest string, configSize int64,
-	layers []ispec.Descriptor, repoBlob2Size map[string]int64,
-) int64 {
-	imgSize := int64(0)
-
-	// add config size
-	imgSize += configSize
-	repoBlob2Size[configDigest] = configSize
-
-	// add manifest size
-	imgSize += manifestSize
-	repoBlob2Size[manifestDigest] = manifestSize
-
-	// add layers size
-	for _, layer := range layers {
-		repoBlob2Size[layer.Digest.String()] = layer.Size
-		imgSize += layer.Size
-	}
-
-	return imgSize
 }
 
 // calcalculateImageMatchingScore iterated from the index of the matched string in the
@@ -881,7 +346,7 @@ func (r *queryResolver) getImageList(store storage.ImageStore, imageName string)
 				}
 
 				isSigned := layoutUtils.CheckManifestSignature(repo, digest)
-				imageInfo := BuildImageInfo(repo, tag.Name, digest, manifest,
+				imageInfo := convert.BuildImageInfo(repo, tag.Name, digest, manifest,
 					imageConfig, isSigned)
 
 				results = append(results, imageInfo)
@@ -894,165 +359,6 @@ func (r *queryResolver) getImageList(store storage.ImageStore, imageName string)
 	}
 
 	return results, nil
-}
-
-func BuildImageInfo(repo string, tag string, manifestDigest godigest.Digest,
-	manifest v1.Manifest, imageConfig ispec.Image, isSigned bool,
-) *gql_generated.ImageSummary {
-	layers := []*gql_generated.LayerSummary{}
-	size := int64(0)
-	log := log.NewLogger("debug", "")
-	allHistory := []*gql_generated.LayerHistory{}
-	formattedManifestDigest := manifestDigest.Hex()
-	annotations := common.GetAnnotations(manifest.Annotations, imageConfig.Config.Labels)
-	lastUpdated := common.GetImageLastUpdated(imageConfig)
-
-	history := imageConfig.History
-	if len(history) == 0 {
-		for _, layer := range manifest.Layers {
-			size += layer.Size
-			digest := layer.Digest.Hex
-			layerSize := strconv.FormatInt(layer.Size, 10)
-
-			layer := &gql_generated.LayerSummary{
-				Size:   &layerSize,
-				Digest: &digest,
-			}
-
-			layers = append(
-				layers,
-				layer,
-			)
-
-			allHistory = append(allHistory, &gql_generated.LayerHistory{
-				Layer:              layer,
-				HistoryDescription: &gql_generated.HistoryDescription{},
-			})
-		}
-
-		formattedSize := strconv.FormatInt(size, 10)
-
-		imageInfo := &gql_generated.ImageSummary{
-			RepoName:      &repo,
-			Tag:           &tag,
-			Digest:        &formattedManifestDigest,
-			ConfigDigest:  &manifest.Config.Digest.Hex,
-			Size:          &formattedSize,
-			Layers:        layers,
-			History:       allHistory,
-			Vendor:        &annotations.Vendor,
-			Description:   &annotations.Description,
-			Title:         &annotations.Title,
-			Documentation: &annotations.Documentation,
-			Licenses:      &annotations.Licenses,
-			Labels:        &annotations.Labels,
-			Source:        &annotations.Source,
-			LastUpdated:   &lastUpdated,
-			IsSigned:      &isSigned,
-			Platform: &gql_generated.OsArch{
-				Os:   &imageConfig.OS,
-				Arch: &imageConfig.Architecture,
-			},
-			Logo: &annotations.Logo,
-		}
-
-		return imageInfo
-	}
-
-	// iterator over manifest layers
-	var layersIterator int
-	// since we are appending pointers, it is important to iterate with an index over slice
-	for i := range history {
-		allHistory = append(allHistory, &gql_generated.LayerHistory{
-			HistoryDescription: &gql_generated.HistoryDescription{
-				Created:    history[i].Created,
-				CreatedBy:  &history[i].CreatedBy,
-				Author:     &history[i].Author,
-				Comment:    &history[i].Comment,
-				EmptyLayer: &history[i].EmptyLayer,
-			},
-		})
-
-		if history[i].EmptyLayer {
-			continue
-		}
-
-		if layersIterator+1 > len(manifest.Layers) {
-			formattedSize := strconv.FormatInt(size, 10)
-
-			log.Error().Err(zerr.ErrBadLayerCount).Msg("error on creating layer history for ImageSummary")
-
-			return &gql_generated.ImageSummary{
-				RepoName:      &repo,
-				Tag:           &tag,
-				Digest:        &formattedManifestDigest,
-				ConfigDigest:  &manifest.Config.Digest.Hex,
-				Size:          &formattedSize,
-				Layers:        layers,
-				History:       allHistory,
-				Vendor:        &annotations.Vendor,
-				Description:   &annotations.Description,
-				Title:         &annotations.Title,
-				Documentation: &annotations.Documentation,
-				Licenses:      &annotations.Licenses,
-				Labels:        &annotations.Labels,
-				Source:        &annotations.Source,
-				LastUpdated:   &lastUpdated,
-				IsSigned:      &isSigned,
-				Platform: &gql_generated.OsArch{
-					Os:   &imageConfig.OS,
-					Arch: &imageConfig.Architecture,
-				},
-				Logo: &annotations.Logo,
-			}
-		}
-
-		size += manifest.Layers[layersIterator].Size
-		digest := manifest.Layers[layersIterator].Digest.Hex
-		layerSize := strconv.FormatInt(manifest.Layers[layersIterator].Size, 10)
-
-		layer := &gql_generated.LayerSummary{
-			Size:   &layerSize,
-			Digest: &digest,
-		}
-
-		layers = append(
-			layers,
-			layer,
-		)
-
-		allHistory[i].Layer = layer
-
-		layersIterator++
-	}
-
-	formattedSize := strconv.FormatInt(size, 10)
-
-	imageInfo := &gql_generated.ImageSummary{
-		RepoName:      &repo,
-		Tag:           &tag,
-		Digest:        &formattedManifestDigest,
-		ConfigDigest:  &manifest.Config.Digest.Hex,
-		Size:          &formattedSize,
-		Layers:        layers,
-		History:       allHistory,
-		Vendor:        &annotations.Vendor,
-		Description:   &annotations.Description,
-		Title:         &annotations.Title,
-		Documentation: &annotations.Documentation,
-		Licenses:      &annotations.Licenses,
-		Labels:        &annotations.Labels,
-		Source:        &annotations.Source,
-		LastUpdated:   &lastUpdated,
-		IsSigned:      &isSigned,
-		Platform: &gql_generated.OsArch{
-			Os:   &imageConfig.OS,
-			Arch: &imageConfig.Architecture,
-		},
-		Logo: &annotations.Logo,
-	}
-
-	return imageInfo
 }
 
 // returns either a user has or not rights on 'repository'.
